Fixed counting of Vecslot.
Also fixed some line lengths.

+++ src/FciMCData.F90
@@ -43,7 +43,7 @@ MODULE FciMCData
       REAL(dp) :: SumSigns, SumSpawns, AvNoatHF
       LOGICAL :: tFillingStochRDMonFly, tFillingExplicRDMonFly
       integer :: Spawned_Parts_Zero, HFInd, NCurrH, IterRDMStart, IterRDM_HF
-      integer, dimension(lenof_sign) :: InstNoatHf
+      real(dp), dimension(lenof_sign) :: InstNoatHf
 
       INTEGER(KIND=n_int) , ALLOCATABLE :: TempSpawnedParts(:,:)
       INTEGER :: TempSpawnedPartsTag, TempSpawnedPartsInd
